requirements.txt
openai
langchain
streamlit
python-dotenv
requests